backend/requirements.txt
Flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
google-generativeai==0.4.0
sentence-transformers==2.2.2
torch==2.6.0
transformers==4.37.2
numpy==1.26.4
scipy==1.12.0
xmltodict==0.13.0
